+++ neoforge/src/main/java/de/cech12/usefulhats/client/UsefulHatsNeoForgeClientEvents.java
@@ -1,6 +1,7 @@
 package de.cech12.usefulhats.client;
 
 import de.cech12.usefulhats.Constants;
+import de.cech12.usefulhats.client.compat.AccessoriesClientCompat;
 import de.cech12.usefulhats.platform.Services;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.geom.LayerDefinitions;
@@ -39,9 +40,9 @@ public class UsefulHatsNeoForgeClientEvents {
 
     @SubscribeEvent
     public static void onClientInit(final FMLClientSetupEvent event) {
-        //if (Services.PLATFORM.isModLoaded(Constants.ACCESSORIES_MOD_ID)) {
-        //    AccessoriesClientCompat.register();
-        //}
+        if (Services.PLATFORM.isModLoaded(Constants.ACCESSORIES_MOD_ID)) {
+            AccessoriesClientCompat.register();
+        }
     }
 
 }